feat: マークダウンは改行文字で即時改行する
resolve #407
