+++ package.json
@@ -61,6 +61,7 @@
     "react-infinite-scroll-hook": "^3.0.0",
     "react-intersection-observer": "^8.32.0",
     "react-markdown": "^6.0.2",
+    "remark-breaks": "^2.0.2",
     "remark-gfm": "^1.0.0",
     "strict-event-emitter-types": "^2.0.0",
     "strip-markdown": "^4.0.0",
